requirements.txt
aiohttp
discord
nest_asyncio
bs4
requests
Pillow
sqlalchemy
business-rule-engine
pyodbc
python-dotenv
pandas
tabulate